src/ai_assistant/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
openai>=1.0.0
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
plotly>=5.18.0
python-dotenv>=1.0.0
seaborn>=0.12.0
scanpy>=1.9.0
anndata>=0.9.0
scipy>=1.11.0 